HashCode2014Sol Documentation

Documentation for HashCode2014Sol.

HashCode2014Sol.Problem

HashCode2014Sol.add_route!

HashCode2014Sol.change_duration

HashCode2014Sol.check_junction

HashCode2014Sol.check_route

HashCode2014Sol.city

HashCode2014Sol.delete_route!

HashCode2014Sol.distance

HashCode2014Sol.get_neighbor_streets

HashCode2014Sol.optimal_neighbor

HashCode2014Sol.optimal_walk

HashCode2014Sol.random_walk_distance

HashCode2014Sol.replace_route!

HashCode2014Sol.route

HashCode2014Sol.routegrid

HashCode2014Sol.routegrid

HashCode2014Sol.upper_limit

HashCode2014Sol.Problem — Type

Problem

Store a Problem that stores a custom default city from HashCode2014 as 'problem', only modifying it by the 'duration' given at construction, as well as its most optimal walk 'solution' and the unique 'distance' traveled in this 'solution' in meters.

Fields

'problem::City': a problem city description

'solution:Solution': a solution to the 'problem' based on the optimal_walk algorithm

'distance::Int': a record, in meters, of the unique distance traveled by the cars in optimal 'solution'

Constructor

Problem(duration::Int)
Returns a Problem instance about a default city with the given 'duration' (in seconds)

HashCode2014Sol.add_route! — Method

add_route!(route_grid, route)

Register a new Route to RouteGrid

HashCode2014Sol.change_duration — Method

change_duration(city, total_duration)

Create a new City with a different total_duration and everything else equal.

HashCode2014Sol.check_junction — Method

check_junction(route_grid, street_idx_1, street_idx_2)

Check if two Streets of the given indices in the RouteGrid intersect at a Junction. If so, return (true, Junction); else, return (false, nothing)

HashCode2014Sol.check_route — Method

check_route(route_grid, junction_idx_start, junction_idx_end)

Check if there exists a Route from two Junctions of the given indices in a RouteGrid. The first index denotes the start junction, and the second index denotes the end junction. If so, return (true, Route). Else, return (false, nothing).

HashCode2014Sol.city — Method

city()

Returns default city of type City in HashCode2014 starter

HashCode2014Sol.delete_route! — Method

delete_route!(route_grid, junc_idx_start, junc_idx_end)

Delete an existing Route between two streets from RouteGrid. The first index denotes the start street, and the second index denotes the end street.

HashCode2014Sol.distance — Method

distance(solution, city)

For a city with a feasible solution, returns the unique distance traveled by the cars

HashCode2014Sol.get_neighbor_streets — Method

city: city for which we want to map neighboring streets for for each street
returns a vector neighbors_streets, such that for every junction with index i: every junction index
    in neighbors_streets[i] can be traveled to from junction i
    Goes even more specific that if street in neighbors_streets[i]:
            street.endpointA = i

HashCode2014Sol.optimal_neighbor — Method

optimal_neighbor(neighbors_streets, visited)

Pre-requisite: all street.endpointA in neighborsstreets is the same Parameters: neighborsstreets: vector of adjacent streets, assumes for all neighbor in neighborsstreets, neighbor.endpointB is travelable from same origin node neighbor.endpointA visited: for all junctions in the corresponding city, a dictionary keeping track of the frequency of how many times each junction was visited (by any car) Returns the street in neighborsstreets that is least visited, breaks ties by picking left-most entry in neighbors_streets

HashCode2014Sol.optimal_walk — Method

optimal_walk(city)

Takes a city, and returns an itinerary for each car that reduces repeated travel along visited streets as much as possible.

Algorithmic description - Modifies randomwalk to avoid visiting already visited junctions as much as possible - To avoid repeated calculation, gets a matrix of all the streets travelable from a junction, switches endpointA and endpointB when bidirectional to put street in the entry of neighborsstreets[endpointB] - Initialize a dictionary visitednodes that keeps track of how many times a junction has been visited, by default: 0 times - For each car: - Get currentnode's travelable streets - Filter streets by which ones match the totalduration - Filter further with optimalneighbor that picks the least visited neighbor - If no applicable streets, move to next car - Add the street's endpointB to current car's itinerary, update visited and duration - Apply Solution(_) to array of all the cars' itineraries in order - Return that Solution

HashCode2014Sol.random_walk_distance — Method

random_walk_distance()

Performs a random walk on the default city() and returns the unique distance traveled by the cars

HashCode2014Sol.replace_route! — Method

replace_route!(route_grid, new_route)

Replace an existing Route between two junctions in RouteGrid with a new Route

HashCode2014Sol.route — Method

route(streets, start_junc_idx, end_junc_idx)

Create a Route from the in-RouteGrid indices of its start and end junctions

HashCode2014Sol.routegrid — Method

routegrid(city)

Create a RouteGrid from a specified City.

HashCode2014Sol.routegrid — Method

routegrid()

Create a RouteGrid from the official challenge City.

HashCode2014Sol.upper_limit — Method

upper_limit(city)

Takes a city and gives an approximate upper limit on the distance that can be feasibly traveled by the cars Algorithmic description: Assumes that all streets are travelable from one another, and that it is actually one car traveling for a limit total of totalduration*nbcars seconds. In that case, we can greedily assume that this one car travels along only the most "efficient" streets, or streets that maximize the meter distance traveled per second. We sort the streets by this efficiency definition and make the car travel along the most efficient, making sure it does not exceed (totalduration*nbcars), and keeping track of the total distance being traveled. At the end, we return this final distance.