HashCode2014Sol Documentation

Documentation for HashCode2014Sol.

HashCode2014Sol.add_route!

HashCode2014Sol.change_duration

HashCode2014Sol.check_junction

HashCode2014Sol.check_route

HashCode2014Sol.delete_route!

HashCode2014Sol.replace_route!

HashCode2014Sol.routegrid

HashCode2014Sol.routegrid

HashCode2014Sol.add_route! — Method

add_route!(route_grid, route)

Register a new Route to RouteGrid

HashCode2014Sol.change_duration — Method

change_duration(city, total_duration)

Create a new City with a different total_duration and everything else equal.

HashCode2014Sol.check_junction — Method

check_junction(route_grid, street_idx_1, street_idx_2)

Check if two Streets of the given indices in the RouteGrid intersect at a Junction. If so, return (true, Junction); else, return (false, nothing)

HashCode2014Sol.check_route — Method

check_route(route_grid, junction_idx_start, junction_idx_end)

Check if there exists a Route from two Junctions of the given indices in a RouteGrid. The first index denotes the start junction, and the second index denotes the end junction. If so, return (true, Route). Else, return (false, nothing).

HashCode2014Sol.delete_route! — Method

delete_route!(route_grid, street_idx_start, street_idx_end)

Delete an existing Route between two streets from RouteGrid. The first index denotes the start street, and the second index denotes the end street.

HashCode2014Sol.replace_route! — Method

replace_route!(route_grid, new_route)

Replace an existing Route between two streets in RouteGrid with a new Route

HashCode2014Sol.routegrid — Method

routegrid(city)

Create a RouteGrid from a specified City.

HashCode2014Sol.routegrid — Method

routegrid()

Create a RouteGrid from the official challenge City.