HashCode2014Sol Documentation

Documentation for HashCode2014Sol.

HashCode2014Sol.Problem

HashCode2014Sol.add_junction_to_route!

HashCode2014Sol.change_duration

HashCode2014Sol.check_junction

HashCode2014Sol.city

HashCode2014Sol.distance

HashCode2014Sol.get_junction_connections

HashCode2014Sol.get_last_junction

HashCode2014Sol.optimal_neighbor

HashCode2014Sol.optimal_walk

HashCode2014Sol.random_walk_distance

HashCode2014Sol.routegrid

HashCode2014Sol.routegrid

HashCode2014Sol.upper_limit

HashCode2014Sol.Problem — Type

Problem

Store a Problem that stores a custom default city from HashCode2014 as problem, only modifying it by the duration given at construction, as well as its most optimal walk solution and the unique distance traveled in this solution in meters.

Fields

problem::City: a problem city description

solution:Solution: a solution to the problem based on the optimal_walk algorithm

distance::Int: a record, in meters, of the unique distance traveled by the cars in optimal solution

Problem(duration::Int)

Parameter

duration::Int - duration in seconds that we want to ascribe to default HashCode2014 city

Constructor

Returns a Problem instance about a default city with the given duration (in seconds)

HashCode2014Sol.add_junction_to_route! — Method

add_junction_to_route!(route_grid, car_id, junction_id)

Expand route of car and increment junction visits

Parameter

route_grid:RouteGrid - our target RouteGrid instance

car_id - the number of the car whose itinerary in route_grid we would like to check

Returns nothing

HashCode2014Sol.change_duration — Method

change_duration(city, total_duration)

Create a new City with a different total_duration and everything else equal.

Parameter

city::City - city whose duration we would like to change

total_duration::Int - the new total_duration (in meters) we would like to change to

Returns a new City instance that changes the total_duration to that specified

HashCode2014Sol.check_junction — Method

check_junction(route_grid, street_idx_1, street_idx_2)

Check if two given treets in a RouteGrid intersect at a junction.

Parameter

route_grid::RouteGrid - the RouteGrid instance the two streets are part of

street_idx_1::Int - the first ordered street #

street_idx_2::Int - the other ordered street # we want to compare the first two

Return

a tuple of form (::Bool, HashCode2014.Junction)

The first value represents whether the two specified streets intersect at a junction

The second value representing if yes, at which junction they intersect; is left null if no

HashCode2014Sol.city — Method

city()

Allows user to create a default HashCode2014 city

Parameter: none

Returns default city of type City in HashCode2014 starter

HashCode2014Sol.distance — Method

distance(solution, city)

For a city with a feasible solution, find the distance traveled in the solution

Parameter

solution::Solution - a feasibility itinerary solution to the junctions the cars travel to

city::City - city whose data values derive the solution

Returns an int representing the unique distance traveled by the cars in the solution in meters

HashCode2014Sol.get_junction_connections — Method

get_junction_connections(streets, junctions)

Get the connecting travelable streets for each junction by index

Parameter

streets::Vector{Street} - all the streets in our target city

junctions::Vector{Junction} - all the junctions in our target city

Returns

A dictionary junction_connections mapping each junction (by its index) to a vector of streets that are travelable from that junction

Goes even more specific that if a street in streets is street.bidirectional:

junction_connections[street.endpointB] contains a new Street instance with the endpoints switched

HashCode2014Sol.get_last_junction — Method

get_last_junction(route_grid, car_id)

Get the last junction that the car has passed in the routegrid.

Parameter

route_grid::RouteGrid - our target RouteGrid instance

car_id - the number of the car whose itinerary in route_grid we would like to check

Returns index of last junction in route_grid visited by car with car_id

HashCode2014Sol.optimal_neighbor — Method

optimal_neighbor(route_grid, query_streets)

Pre-requisite: all street.endpointA in querystreets is the same Parameters: routegrid: the current RouteGrid in context querystreets: vector of adjacent streets, assumes for all neighbor in querystreets, neighbor.endpointB is travelable from same origin node neighbor.endpointA Returns the street in querystreets that is least visited, breaks ties by picking left-most entry in querystreets

HashCode2014Sol.optimal_walk — Method

optimal_walk(city)

For a city, get a solution that aims to prioritize undiscovered/infrequent junctions in its cars' traversals

Parameter

city::City - city whose optimal solution we would like to compute

Returns an itinerary for each car that reduces repeated travel along visited streets as much as possible.

Algorithmic description in documentation

HashCode2014Sol.random_walk_distance — Method

random_walk_distance()

Performs a random walk on the default city() and gets the distance

Parameter: none

Returns an int representing the unique distance traveled by the cars in the solution in meters

HashCode2014Sol.routegrid — Method

routegrid(city)

Create a RouteGrid from a specified City.

Parameter

city::City - City instance we want to derive a RouteGrid from

Returns a new RouteGrid instance based on the data of city

HashCode2014Sol.routegrid — Method

routegrid()

Create a RouteGrid from the official challenge City. Parameter: none Returns a default RouteGrid instance based on a default HashCode2014.City instance

HashCode2014Sol.upper_limit — Method

upper_limit(city)

Calculate an approximate upper limit on the distance travelable in city

Parameter

city::City: target City instance

Returns an integer, in meters, representing the upper limit travelable by cars in city

Algorithm in documentation